# swagger/openapi .json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Online Store - OpenAPI 3.0",
    "description": "This is a backend for online store and implements:\n  - Product CRUD (include option and variants)\n  - Category CRUD\n  - Cart CRUD\n  - Order CRUD\n  \nSome useful links:\n- [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)\n- [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "https://host:port/api/v1"
    }
  ],
...
  "paths": {
...
    "/order": {
      "post": {
        "tags": [
          "order"
        ],
        "summary": "Create order",
        "description": "Create order with form data and delete ordered items from cart",
        "operationId": "createOrder",
        "requestBody": {
          "description": "Create product object",
          "content": {
            "application/form-data": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found"
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
    "/order/{orderId}": {
      "get": {
        "tags": [
          "order"
        ],
        "summary": "Find order by ID",
        "description": "Returns a single order",
        "operationId": "getOrderById",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "order"
        ],
        "summary": "Update a order with form data",
        "operationId": "updateOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of sku",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Create sku object",
          "content": {
            "application/form-data": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "order"
        ],
        "summary": "Delete order by Id",
        "operationId": "deleleteOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "order Id that need to be considered for filter",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
    "/admin/order/{orderId}": {
      "put": {
        "tags": [
          "order"
        ],
        "summary": "Update order status",
        "description": "",
        "operationId": "updateOrderStatus",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of order",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "status of order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrderResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
          },
          "userId": {
            "type": "integer",
            "format": "int",
            "example": 3
          },
          "address": {
            "type": "string",
            "example": "Some Street 111/2"
          },
          "phone": {
            "type": "string",
            "example": "+99300000"
          },
          "comment": {
            "type": "string",
            "example": "here you comment"
          },
          "status": {
            "type": "string",
            "example": "Sended"
          },
          "orderItems": {
            "$ref": "#/components/schemas/OrderItems"
          }
        }
      },
      "OrderRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int",
            "example": 3
          },
          "address": {
            "type": "string",
            "example": "Some Street 111/2"
          },
          "phone": {
            "type": "string",
            "example": "+99300000"
          },
          "comment": {
            "type": "string",
            "example": "here you comment"
          },
          "notes": {
            "type": "string",
            "example": "here you note"
          }
        }
      },
...
      "OrderItems": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int",
            "example": 3
          },
          "skuId": {
            "type": "integer",
            "format": "int",
            "example": 3
          },
          "codename": {
            "type": "string",
            "example": "sku-123456"
          },
          "quantity": {
            "type": "string",
            "example": 2
          },
          "price": {
            "type": "string",
            "example": 12.3
          },
          "subTotal": {
            "type": "string",
            "example": "Sended"
          },
          "skuImage": {
            "type": "string",
            "example": "/static/sku-123456/image1.jpg"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "token_key": {
        "type": "apiKey",
        "name": "token",
        "in": "header"
      }
    }
  }
}